mt7610: fix TxPower delta printing corner case
Delta value specified in 0.5dBm steps, so we should print this
fractional part in some way.

If we independently calculate integer and fractional part like:
  printf("... %d.%d", val / 2, val & 1 ? 5 : 0)
then we have corner case with -1 value, which should be printed as -0.5.
But due to the fact that zero have no sign (e.g. -1/2 = 0) we got +0.5
instead.

So treat value as floating point, divide it by two and print as floating
point value. This is not so effective as spliting on to integer and
fractional parts, but makes code stable and simple, which is actually
important for such type of software as data parsers.

Rework other power printing code in the same way, even they are not
dealing with negative value. Just to make code smooth.