Fix issue with special section alignment
Structures that are placed to the special section (e.g. chip
descriptors) will be padded by a linker. We could use a trick with
substracting location of sturctures with known positions. But with the
LTO compiler options structures could be randomly ordered and that
breaks any our assumptions about structure positions.

To overcome all this hardness utilize trick from the latest iw utility
and place only chip structure pointers to the '__chip' special section
and ignore any zero padding elements.