usb: add specific device selection support
It is not unusual to have multiple USB adapters connected to a host, or
even have a device with known supported chip, but with unknown VID/PID
identifiers. In these cases work with a first known device is not a
good approach.

Add support for specifing USB device as an argument to the USB connector
command line option ('-U'). Device specification acts like a filter,
which narrows devices rage. It is possible to specify item device by its
address or to specify device connection port to be shure that the
utility will interact only with a specific device. Moverover selector
allows to force the utility to work with unknown device by explicitly
specify VID/PID.

To keep existing behaviour, special selector 'any' provided. If user
specify 'any' as a command line option argument, then the utility as
earlier will select a first USB device with known pair of VID/PID.